adding new functions
